Define details for saving and making events online (#893)
