+++ open_event/views/admin/models_views/events.py
@@ -5,28 +5,12 @@ from flask_admin import expose
 from flask_admin.contrib.sqla import ModelView
 from flask.ext import login
 
+from open_event.helpers.helpers import fields_not_empty, string_empty
 from ....helpers.data import DataManager, save_to_db
 from ....helpers.data_getter import DataGetter
 import datetime
 from werkzeug.utils import secure_filename
 from werkzeug.datastructures import ImmutableMultiDict
-
-def string_empty(string):
-    if type(string) is not str and type(string) is not unicode:
-        return False
-    if string and string.strip() and string != u'' and string != u' ':
-        return False
-    else:
-        return True
-
-def string_not_empty(string):
-    return not string_empty(string)
-
-def fields_not_empty(obj, fields):
-    for field in fields:
-        if string_empty(getattr(obj, field)):
-            return False
-    return True
 
 class EventsView(ModelView):
     def is_accessible(self):
@@ -55,6 +39,9 @@ class EventsView(ModelView):
                     filename = secure_filename(img_file.filename)
                     img_file.save(os.path.join(os.path.realpath('.') + '/static/media/image/', filename))
             event = DataManager.create_event(request.form, imd)
+            if request.form.get('state', u'Draft') == u'Published' and string_empty(event.location_name):
+                flash("Your event was saved. To publish your event please review the highlighted fields below.", "warning")
+                return redirect(url_for('.edit_view', event_id=event.id) + "#step=location_name")
             if event:
                 return redirect(url_for('.details_view', event_id=event.id))
             return redirect(url_for('.index_view'))
@@ -69,7 +56,8 @@ class EventsView(ModelView):
 
         checklist = {"": ""}
 
-        if fields_not_empty(event, ['name', 'start_time', 'end_time', 'location_name', 'organizer_name', 'organizer_description']):
+        if fields_not_empty(event, ['name', 'start_time', 'end_time', 'location_name', 'organizer_name',
+                                    'organizer_description']):
             checklist["1"] = 'success'
         elif fields_not_empty(event, ['name', 'start_time', 'end_time']):
             checklist["1"] = 'missing_some'
@@ -147,6 +135,10 @@ class EventsView(ModelView):
         if request.method == "POST":
             event = DataManager.edit_event(request, event_id, event, session_types, tracks, social_links,
                                            microlocations, call_for_speakers, sponsors)
+            if request.form.get('state', u'Draft') == u'Published' and string_empty(event.location_name):
+                flash("Your event was saved. To publish your event please review the highlighted fields below.",
+                      "warning")
+                return redirect(url_for('.edit_view', event_id=event.id) + "#step=location_name")
             return redirect(url_for('.details_view', event_id=event_id))
 
     @expose('/<event_id>/delete/', methods=('GET',))
